aosc-mklive.sh: implement memory detection
This enables LiveKit to boot into different mode depends on how much RAM
is available to GRUB - not the actual amount of physical RAM.

For Retro targets, LiveKit will not be able to boot if RAM requirement
is not met.

Since an all-in-one config file is used, only toggling a value is
required to switch modes.