scripts/livekit.sh: do not use host system dracut configuration
On MIPS, we ship a minimal dracut.conf to make sure that initramfs can be
smaller. However, this seems to have caused LiveKit to fail to mount
/dev/root (one of the omissions have caused this).

This is really a workaround, as we have yet to find the root cause to this
issue. As we are relying on the firmware/GRUB to not to fail to allocate
enough memory for the boot payload, this may still cause the boot to fail.

We will return to this on a later date.