scripts: create an empty machine-id for systemd
Creating an empty machine-id file for systemd to fill it on the first
boot, since we are now defaulting to read-only for the boot process,
and once the filesystem is read only, files can not be created.
Systemd bind-mounts a temporary file to /etc/machine-id and updates
it once the filesystem is mounted read-write. Thus, an empty machine-
id file is required for systemd to function properly.